Implement user-defined order, ensure a 500 in original url does not break the flow
